refactor: move exhibit data to app config
