+++ app/visit/page.tsx
@@ -2,45 +2,8 @@
 import React from 'react';
 import { CardGroup, Grid, GridContainer } from '@trussworks/react-uswds';
 import { Paragraph } from 'app/components/common/Paragraph';
-import ExhibitCard, {
-  Exhibit,
-} from 'app/components/common/cards/small/exhibit';
-
-export const exhibits: Exhibit[] = [
-  {
-    id: 'nasa-hq',
-    heading: 'NASA HQ',
-    openingHours: ['Monday - Friday', '8:30am - 5:30pm'],
-    address: [
-      'NASA Headquarters (East Lobby)',
-      '300 E St. SW, Washington, DC 20546',
-    ],
-    imgSrc: '/images/visit/nasa-hq.png',
-    imgAlt: 'NASA HQ exhibit view',
-  },
-  {
-    id: 'smithsonian-museum',
-    heading: 'Smithsonian National Museum of Natural History',
-    openingHours: ['Every day (closed on Christmas)', '10am - 5:30pm'],
-    address: [
-      'Smithsonian National Museum of Natural History',
-      '1000 Madison Dr, Washington, DC 20560',
-    ],
-    imgSrc: '/images/visit/smithsoniam-museum.jpg',
-    imgAlt: 'Smithsonian National Museum of Natural History exterior view',
-  },
-  {
-    id: 'kennedy-space-center',
-    heading: 'Kennedy Space Center',
-    openingHours: ['Every day (closed on Christmas)', '9am - 5pm'],
-    address: [
-      'KSC Visitor Complex (The Gantry at LC-39)',
-      'Space Commerce Way, Merritt Island, FL 32953',
-    ],
-    imgSrc: '/images/visit/kennedy-space-center.png',
-    imgAlt: 'Kennedy Space Center exterior view',
-  },
-];
+import ExhibitCard from 'app/components/common/cards/small/exhibit';
+import { exhibits } from 'app/config';
 
 const VisitPage: React.FC = () => {
   return (
